requirements.txt
Flask>=1.1
pydantic>=1.6
PyYAML>=5.3
bcrypt>=3.1.7
python-dotenv>=0.14.0
requests>=2.24.0
SQLAlchemy>=1.3.18
Flask-SQLAlchemy>=2.4.4
